refactor(mahamantra): Float → Integer conversion - COSMIC_FRAME scaling (Phase 2)
All resonance/coherence/integrity values now use integer scaling:
- 21600 = 100% (COSMIC_FRAME = arc-minutes in circle)
- 0 = 0%

Files converted:
- cell.py: integrity (0-21600)
- byte.py: coherence, stability, resonance_check (0-21600)
- yajna.py: coherence (0-21600)
- cluster.py: coherence (0-28800 with MALA bonus)
- nadi.py: resonance_ratio, mala_progress (0-21600)
- harmonics.py: all threshold methods accept int resonance
- resonator.py: resonance_strength (0-21600)
- protocol.py: get_resonance (0-21600)

Pattern: Internal math uses float, return value scaled to int.
SSOT: COSMIC_FRAME (21600) from protocols/_seed.py

+++ vibe_core/mahamantra/substrate/cluster.py
@@ -48,7 +48,7 @@ class MahaCluster(Generic[C]):
     
     cells: List[MahaCellUnified[C]]
     resonance_attractor: int        # Common attractor (18, 22, 49, 87, 136)
-    coherence: float                # 0.0 - 1.333+ (MALA-normalized)
+    coherence: int                  # 0 - 28800 (COSMIC_FRAME + 33% MALA bonus)
     
     # Meta-header generated post-init
     cluster_header: MahaHeader = field(init=False)
